# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CloudsWork Solar Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/deposit:
    post:
      operationId: createDeposit
      tags: [wallet]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: Deposit submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /wallet/withdraw:
    post:
      operationId: createWithdrawal
      tags: [wallet]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "200":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /wallet/transactions:
    get:
      operationId: getTransactions
      tags: [wallet]
      summary: Get transaction history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, deposit, withdrawal, profit, commission, purchase, refund]
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
...
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletion"

  # NOTIFICATIONS
...
  /admin/tasks:
    get:
      operationId: getAdminTasks
      tags: [admin]
      summary: List all tasks (admin)
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [admin]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /admin/tasks/{id}:
    put:
      operationId: updateTask
      tags: [admin]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [admin]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        paymentMethod:
          type: string
        voucherUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, amount, currency, paymentMethod, status, createdAt, updatedAt]
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        walletType:
          type: string
        accountTitle:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, amount, fee, netAmount, walletType, status, createdAt, updatedAt]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, profit, commission, purchase, refund, bonus]
        amount:
          type: number
        description:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        createdAt:
          type: string
      required: [id, userId, type, amount, description, status, createdAt]
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [transactions, total, page, totalPages]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [check_in, watch_video, visit_website, share_link]
        reward:
          type: number
        link:
          type: ["string", "null"]
        isCompleted:
          type: boolean
        isActive:
          type: boolean
      required: [id, title, description, type, reward, isCompleted, isActive]
    TaskCompletion:
      type: object
      properties:
        success:
          type: boolean
        reward:
          type: number
        message:
          type: string
      required: [success, reward, message]
...
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        paymentMethod:
          type: string
        voucherUrl:
          type: ["string", "null"]
      required: [amount, currency, paymentMethod]
    WithdrawalInput:
      type: object
      properties:
        amount:
          type: number
        walletType:
          type: string
        accountTitle:
          type: ["string", "null"]
        iban:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
      required: [amount, walletType]
...
    TaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [check_in, watch_video, visit_website, share_link]
        reward:
          type: number
        link:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [title, description, type, reward, isActive]
...